chore: moved components except App.vue in same directory.
