+++ web-frontend/src/router/index.js
@@ -1,8 +1,8 @@
 import { createRouter, createWebHistory } from 'vue-router'
 
-import Home from '../Home.vue'
-import About from '../About.vue'
-import Works from '../Works.vue'
+import Home from '../components/Home.vue'
+import About from '../components/About.vue'
+import Works from '../components/Works.vue'
 
 const routes = [
   {
